policy: Refactor L4PolicyMap to Index Port Ranges
The L4PolicyMap needs to index L4Filters by port
and protocol. With the introduction of port ranges
the golang builtin map type no longer suffices to
index the L4Filters. This commit refactors the
type to take a mix of named ports (which can never
have an EndPort) as well as simple port numbers,
and ranges. All usage of the type is rewritten
to account for the changes.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>